backend/requirements.txt
Django
drf_yasg
dj_rest_auth
paypalrestsdk
phonenumbers
unidecode
stripe
celery
python-decouple
python-dotenv
django-rest-auth
django-cors-headers
django-allauth
django-phonenumber-field
django_filter
django-import-export
django-paypal
django-countries
djangorestframework-simplejwt
Pillow
django_cloudinary_storage